Added missing cast to unsigned int for bitset index

+++ src/map.h
@@ -261,7 +261,7 @@ class map
         /*@}*/
 
         void set_memory_seen_cache_dirty( const tripoint p ) {
-            get_cache( p.z ).map_memory_seen_cache[ p.x + p.y * MAPSIZE_Y ] = false;
+            get_cache( p.z ).map_memory_seen_cache[ static_cast<size_t>( p.x + ( p.y * MAPSIZE_Y ) ) ] = false;
         }
 
         bool check_and_set_seen_cache( const tripoint &p ) const {
